un bel passo avanti

+++ lib/common/personalCard.dart
@@ -3,12 +3,11 @@ import 'package:flutter/material.dart';
 class PersonalCard {
   final String photo;
   final String name;
-  double amount;
+  int amount;
 
   PersonalCard({required this.photo, required this.name, required this.amount});
 
-  void addAmount(double amount) {
+  void addAmount(int amount) {
     this.amount += amount;
   }
-
 }
